refactor(semantic): centralize function call resolution
